+++ sale-service/src/main/java/vn/viettel/sale/service/dto/ExchangeTransDTO.java
@@ -23,6 +23,6 @@ public class ExchangeTransDTO {
     private Long reasonId;
     private String reason;
     private Integer quantity;
-    private Float totalAmount;
+    private Double totalAmount;
     private List<ExchangeTransDetailRequest> listProducts;
 }
